Convert benchmarks to use bencher and run on stable
