+++ Cargo.toml
@@ -23,3 +23,14 @@ lumol-input = {path = "src/input"}
 lumol-core = {path = "src/core"}
 log = "0.3"
 clap = "2"
+
+[dev-dependencies]
+bencher = "0.1"
+
+[[bench]]
+name = "water"
+harness = false
+
+[[bench]]
+name = "nacl"
+harness = false
