--- benches/results/2016-03-14-23:53-x86_64-apple-darwin.bench
@@ -1,11 +1,0 @@
-# rustc 1.9.0-nightly (6d215fe04 2016-03-14)
-# 2a43eaf 2016-03-14 23:53:41 +0100 Mutable global potentials in energy computation
-     Running target/release/ewald-bcb1ab681fa4b323
-
-running 4 tests
-test nacl::energy  ... bench:   1,636,287 ns/iter (+/- 290,976)
-test nacl::forces  ... bench:  15,953,889 ns/iter (+/- 1,975,519)
-test water::energy ... bench:  28,395,730 ns/iter (+/- 3,899,270)
-test water::forces ... bench: 253,296,275 ns/iter (+/- 26,978,709)
-
-test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured
