--- benches/results/2016-03-24-21:17-x86_64-apple-darwin.bench
@@ -1,12 +1,0 @@
-# rustc 1.9.0-nightly (02310fd31 2016-03-19)
-# 1c41fbe 2016-03-24 21:17:47 +0100 Add a cache for molecules ids
-     Running target/release/ewald-bcb1ab681fa4b323
-
-running 4 tests
-test nacl::energy  ... bench:     371,496 ns/iter (+/- 149,020)
-test nacl::forces  ... bench:  13,702,581 ns/iter (+/- 1,601,409)
-test water::energy ... bench:   2,457,884 ns/iter (+/- 352,650)
-test water::forces ... bench: 211,773,433 ns/iter (+/- 15,257,106)
-
-test result: ok. 0 passed; 0 failed; 0 ignored; 4 measured
-
